chore: process result in common
